feat: 상세 일정 화면에 선택한 팀원들의 이름이 나오도록 구현 (#60)
